feat(cli): use a better naming scheme
